# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WABA Control identity and organization API
  version: 0.2.0
  description: >-
    Implemented Sprint 0 health and Sprint 1 company identity/access endpoints. No public registration or Meta/product
    messaging endpoints. Cookie mutations require matching Origin, CSRF cookie/header and the session-bound CSRF digest.
    Organization IDs never establish authorization.
  license:
    name: Proprietary
    url: https://choosealicense.com/no-permission/
servers:
  - url: /
security: []
...
paths:
...
  /api/v1/auth/csrf:
    get:
      operationId: csrf
      summary: csrf
      tags:
        - Identity
      security: []
      description: Return/initiate the same-origin CSRF cookie and readable token.
      x-permission: Public proof/throttling or CSRF bootstrap
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/v1/auth/login:
    post:
      operationId: public_login
      summary: public login
      tags:
        - Identity
      security: []
      description: >-
        Generic authentication failure. Successful password authentication returns a rotated session or a bounded MFA
        challenge.
      x-permission: Public proof/throttling or CSRF bootstrap
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  maxLength: 256
                  writeOnly: true
                code:
                  type: string
                  writeOnly: true
                token:
                  type: string
                  writeOnly: true
                name:
                  type: string
                  maxLength: 100
                organization_id:
                  type: string
                  format: uuid
                member_id:
                  type: string
                  format: uuid
                role_ids:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    enum:
                      - organization.view
                      - members.view
                      - members.invite
                      - members.manage
                      - teams.view
                      - teams.manage
                      - roles.view
                      - roles.manage
                      - roles.assign
                      - owners.manage
                      - audit.view
                scope:
                  type: string
                  enum:
                    - ORG
                    - TEAM
                    - SELF
                team_id:
                  anyOf:
                    - type: string
                      format: uuid
                    - type: 'null'
              required:
                - email
                - password
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: logout
      tags:
        - Identity
      security:
        - SessionCookie: []
      description: Revoke current session and expire its cookie.
      x-permission: Authenticated self
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 256
                  writeOnly: true
                code:
                  type: string
                  writeOnly: true
                token:
                  type: string
                  writeOnly: true
                name:
                  type: string
                  maxLength: 100
                organization_id:
                  type: string
                  format: uuid
                member_id:
                  type: string
                  format: uuid
                role_ids:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    enum:
                      - organization.view
                      - members.view
                      - members.invite
                      - members.manage
                      - teams.view
                      - teams.manage
                      - roles.view
                      - roles.manage
                      - roles.assign
                      - owners.manage
                      - audit.view
                scope:
                  type: string
                  enum:
                    - ORG
                    - TEAM
                    - SELF
                team_id:
                  anyOf:
                    - type: string
                      format: uuid
                    - type: 'null'
              required: []
  /api/v1/auth/session:
    get:
      operationId: session
      summary: session
      tags:
        - Identity
      security:
        - SessionCookie: []
      description: Return current nonsecret identity metadata and active organization memberships.
      x-permission: Authenticated self
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/v1/auth/mfa/verify:
    post:
      operationId: public_mfa
      summary: public mfa
      tags:
        - Identity
      security: []
      description: Consume an MFA login challenge after one-use TOTP/recovery-code validation; establish a rotated session.
      x-permission: Public proof/throttling or CSRF bootstrap
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 256
                  writeOnly: true
                code:
                  type: string
                  writeOnly: true
                token:
                  type: string
                  writeOnly: true
                name:
                  type: string
                  maxLength: 100
                organization_id:
                  type: string
                  format: uuid
                member_id:
                  type: string
                  format: uuid
                role_ids:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    enum:
                      - organization.view
                      - members.view
                      - members.invite
                      - members.manage
                      - teams.view
                      - teams.manage
                      - roles.view
                      - roles.manage
                      - roles.assign
                      - owners.manage
                      - audit.view
                scope:
                  type: string
                  enum:
                    - ORG
                    - TEAM
                    - SELF
                team_id:
                  anyOf:
                    - type: string
                      format: uuid
                    - type: 'null'
              required:
                - token
                - code
  /api/v1/auth/reauthenticate:
    post:
      operationId: reauth
      summary: reauth
      tags:
        - Identity
      security:
        - SessionCookie: []
      description: Password plus current factor if enrolled. Rotate session; five-minute reauthentication freshness.
      x-permission: Authenticated self
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  maxLength: 256
                  writeOnly: true
                code:
                  type: string
                  writeOnly: true
                token:
                  type: string
                  writeOnly: true
                name:
                  type: string
                  maxLength: 100
                organization_id:
                  type: string
                  format: uuid
                member_id:
                  type: string
                  format: uuid
                role_ids:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    enum:
                      - organization.view
                      - members.view
                      - members.invite
                      - members.manage
                      - teams.view
                      - teams.manage
                      - roles.view
                      - roles.manage
                      - roles.assign
                      - owners.manage
                      - audit.view
                scope:
                  type: string
                  enum:
                    - ORG
                    - TEAM
                    - SELF
                team_id:
                  anyOf:
                    - type: string
                      format: uuid
                    - type: 'null'
              required:
                - password
  /api/v1/auth/password-reset/request:
    post:
      operationId: public_reset_request
      summary: public reset request
      tags:
        - Identity
      security: []
      description: >-
        Generic 202 for known/unknown addresses. Atomically supersede the previous reset challenge and enqueue encrypted
        identity mail for eligible identities.
      x-permission: Public proof/throttling or CSRF bootstrap
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/Origin'
      responses:
        '202':
          description: Accepted for asynchronous email delivery; not proof of inbox receipt.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 256
                  writeOnly: true
                code:
                  type: string
                  writeOnly: true
                token:
                  type: string
                  writeOnly: true
                name:
                  type: string
                  maxLength: 100
                organization_id:
                  type: string
                  format: uuid
                member_id:
                  type: string
                  format: uuid
                role_ids:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    enum:
                      - organization.view
                      - members.view
                      - members.invite
                      - members.manage
                      - teams.view
                      - teams.manage
                      - roles.view
                      - roles.manage
                      - roles.assign
                      - owners.manage
                      - audit.view
                scope:
                  type: string
                  enum:
                    - ORG
                    - TEAM
                    - SELF
                team_id:
                  anyOf:
                    - type: string
                      format: uuid
                    - type: 'null'
              required:
                - email
  /api/v1/auth/password-reset/complete:
    post:
      operationId: public_reset_complete
      summary: public reset complete
      tags:
        - Identity
      security: []
      description: >-
        Consume a single-use 30-minute reset challenge, change the password, invalidate old sessions and enqueue a
        security notice.
      x-permission: Public proof/throttling or CSRF bootstrap
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/Origin'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 256
                  writeOnly: true
                code:
                  type: string
                  writeOnly: true
                token:
                  type: string
                  writeOnly: true
                name:
                  type: string
                  maxLength: 100
                organization_id:
                  type: string
                  format: uuid
                member_id:
                  type: string
                  format: uuid
                role_ids:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    enum:
                      - organization.view
                      - members.view
                      - members.invite
                      - members.manage
                      - teams.view
                      - teams.manage
                      - roles.view
                      - roles.manage
                      - roles.assign
                      - owners.manage
                      - audit.view
                scope:
                  type: string
                  enum:
                    - ORG
                    - TEAM
                    - SELF
                team_id:
                  anyOf:
                    - type: string
                      format: uuid
                    - type: 'null'
              required:
                - token
                - password
...
  /api/v1/security/sessions:
    get:
      operationId: sessions
      summary: sessions
      tags:
        - Identity
      security:
        - SessionCookie: []
      description: List own unrevoked/unexpired sessions; no bearer or CSRF digest.
      x-permission: Authenticated self
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: query
          name: cursor
          required: false
          description: >-
            Opaque signed continuation bound to the authenticated user, organization and endpoint. Reauthorization
            occurs on every page.
          schema:
            type: string
            maxLength: 128
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: waba_session
      description: >-
        Opaque server-side session. HttpOnly; Secure in production; SameSite=Lax; 12-hour absolute and 30-minute idle
        expiry. No raw session secret in JSON.
    APIKey:
      type: http
      scheme: bearer
      description: Future organization/resource-scoped machine key. Not implemented.
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Valid bounded IDs are propagated; missing/invalid values are replaced.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]{1,64}$
    CSRF:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
        minLength: 43
        maxLength: 43
      description: Must match the host-only waba_csrf cookie; authenticated actions also verify its session-stored digest.
    Origin:
      in: header
      name: Origin
      required: true
      schema:
        type: string
      description: Exact configured PUBLIC_ORIGIN. Cross-origin state changes are rejected.
...
  responses:
...
    Error:
      description: Safe error with no underlying credentials or dependency details.
      headers:
        X-Request-ID:
          schema:
            type: string
          description: Server correlation ID.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    IdentityList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            description: >-
              Resource rows contain only the requested organization's records or the authenticated user's sessions; no
              credential columns.
        next_cursor:
          type:
            - string
            - 'null'
        has_more:
          type: boolean
      description: >-
        At most 100 records per page. Follow next_cursor until null; permission and tenant scope are reevaluated on
        every request.